Use new smaller image for bevy_ecs_ldtk (#288)
@Trouv with your permission, I'd like to swap out `bevy_ecs_ldtk`'s image with this smaller version.

You can see it animated side by side with the previous version in the changed files tab.

<img width="311" alt="image" src="https://user-images.githubusercontent.com/200550/213820696-40d1b246-fa25-4feb-8ebc-5b3a4f401576.png">

This is a re-recording that hopefully captures the same spirit and image quality. I took some liberties to match the game's aspect ratio to what's displayed on the website and get the file size down.

Context: I'm trying to add image size validation to the CI (https://github.com/bevyengine/bevy-website/pull/464), and there are a few images that are currently over the size limit.

If you'd prefer to cut a new recording yourself, that's great. There's no rush. I just thought it would be rude to ask you to re-do it without giving you an easy option.